serial: ns16550: separate dt and platform specifics in probing function
The ns16550 probe function mix dt and platform specific code. Rework the
function to separate them. This also has the benefit to fix the
'clock-frequency' dt parameter which cannot be used: if no platform data
was provided (e.g. when using dt), the driver tried to get a clock and
failed which led to return from the probing function with an error.

This rework keep the initial behaviour: the platform data having priority
over the dt properties.

Separating this commit into two clear steps is not simple, the second
not keeping the changes made in the first one. Thus keep a single patch.

Signed-off-by: Antoine Tenart <antoine.tenart@free-electrons.com>